requirements.txt
metaflow
pandas
wget
boto3
python-dotenv
fire
supertriplets
matplotlib